Reworked the class structure to be easier to extend
