+++ FF1Lib/ExtensiveHints/IHintSource.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FF1Lib
+{
+	public interface IHintSource
+	{
+		List<GeneratedHint> GetHints();
+	}
+}
